refactor: switch to being sticky by default
